requirements.txt
colorama==0.4.6
Pillow==10.1.0
pyfiglet==1.0.2
